Add CoreODEObjects and use singelton for the concrete factory
