+++ sim/src/core/NodeManager.h
@@ -37,9 +37,6 @@
 #include <mars/interfaces/sim/ControlCenter.h>
 #include <mars/interfaces/sim/NodeManagerInterface.h>
 
-#include <ObjectFactoryInterface.h>
-#include <ODEObjectFactory.h>
-
 namespace mars {
   namespace sim {
 
@@ -165,8 +162,6 @@ namespace mars {
                         const std::string &value);
 
     private:
-
-      ObjectFactoryInterface* odeObjectFactory; 
 
       interfaces::NodeId next_node_id;
       bool update_all_nodes;
