+++ sim/src/physics/ODEObjectFactory.cpp
@@ -19,13 +19,7 @@
  */
 
 #include <ODEObjectFactory.h>
-#include <ODEBox.h>
-#include <ODECapsule.h>
-#include <ODECylinder.h>
-#include <ODEMesh.h>
-#include <ODEPlane.h>
-#include <ODESphere.h>
-#include <ODEHeightField.h>
+#include <CoreODEObjects.h>
 
 #include <memory>
 
@@ -34,12 +28,10 @@ namespace sim{
 
 using namespace ::mars::interfaces;
 
-ODEObjectFactory* ODEObjectFactory::instance = NULL;
+//ODEObjectFactory* ODEObjectFactory::instance = NULL;
 
-ODEObjectFactory* ODEObjectFactory::getInstance(){
-  if (NULL == instance){
-    instance = new ODEObjectFactory();
-  }
+ODEObjectFactory& ODEObjectFactory::Instance(){
+  static ODEObjectFactory instance;
   return instance;
 } 
 
@@ -52,25 +44,25 @@ ODEObjectFactory::~ODEObjectFactory(){
 std::shared_ptr<NodeInterface> ODEObjectFactory::createObject(std::shared_ptr<PhysicsInterface> worldPhysics, NodeData * nodeData){
   switch(nodeData->physicMode) {
   case NODE_TYPE_BOX:
-    return std::static_pointer_cast<NodeInterface>(std::make_shared<ODEBox>(worldPhysics, nodeData));
+    return std::make_shared<ODEBox>(worldPhysics, nodeData);
     break;
   case NODE_TYPE_CAPSULE:
-    return std::static_pointer_cast<NodeInterface>(std::make_shared<ODECapsule>(worldPhysics, nodeData));
+    return std::make_shared<ODECapsule>(worldPhysics, nodeData);
     break; 
   case NODE_TYPE_CYLINDER:
-    return std::static_pointer_cast<NodeInterface>(std::make_shared<ODECylinder>(worldPhysics, nodeData));
+    return std::make_shared<ODECylinder>(worldPhysics, nodeData);
     break; 
   case NODE_TYPE_MESH:
-    return std::static_pointer_cast<NodeInterface>(std::make_shared<ODEMesh>(worldPhysics, nodeData));
+    return std::make_shared<ODEMesh>(worldPhysics, nodeData);
     break;  
   case NODE_TYPE_PLANE:
-    return std::static_pointer_cast<NodeInterface>(std::make_shared<ODEPlane>(worldPhysics, nodeData));
+    return std::make_shared<ODEPlane>(worldPhysics, nodeData);
     break; 
   case NODE_TYPE_SPHERE:
-    return std::static_pointer_cast<NodeInterface>(std::make_shared<ODESphere>(worldPhysics, nodeData));
+    return std::make_shared<ODESphere>(worldPhysics, nodeData);
     break;    
   case NODE_TYPE_TERRAIN:
-    return std::static_pointer_cast<NodeInterface>(std::make_shared<ODEHeightField>(worldPhysics, nodeData));
+    return std::make_shared<ODEHeightField>(worldPhysics, nodeData);
     break; 
   default:
     // no correct type is spezified, so no physically node will be created
